requirements.txt
aws-cdk.aws-lambda ~= 1.134.0
aws-cdk.aws-eks ~= 1.134
